Revert "pmt-ptw"
This reverts commit c5a05555dadbd71dc817ff6b3c7672e6953daad3.

+++ src/main/java/io/reactivex/rxjava3/internal/disposables/EmptyDisposable.java
@@ -111,7 +111,7 @@ public enum EmptyDisposable implements QueueDisposable<Object> {
 
     @Override
     public int requestFusion(int mode) {
-        return ((Integer)(mode)).intValue() & ASYNC;
+        return mode & ASYNC;
     }
 
 }
